last update

+++ app/Services/CourseService.php
@@ -2,7 +2,10 @@
 
 namespace App\Services;
 
+use App\Http\Resources\CourseResource;
 use App\Models\Course;
+use Illuminate\Http\JsonResponse;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Validator;
 
 class CourseService
@@ -15,8 +18,8 @@ class CourseService
             'files.*' => 'nullable|file|mimes:jpeg,png,jpg,gif,pdf,mp4|max:2048',
             'location' => 'nullable',
             'discount' => 'nullable',
-            'user_id' => 'required|exists:users,id',
         ]);
+        $data['user_id'] = Auth::id();
 
         if ($validator->fails()) {
             return [
@@ -25,15 +28,12 @@ class CourseService
                 'errors' => $validator->errors(),
             ];
         }
-
-        // Handle file upload if needed
-
         $course = Course::create($data);
 
         return [
             'success' => true,
             'message' => 'Course created successfully',
-            'data' => $course,
+            'data' => new CourseResource($course),
         ];
     }
 
@@ -47,6 +47,7 @@ class CourseService
             'discount' => 'nullable',
             'user_id' => 'required|exists:users,id',
         ]);
+        $data['user_id'] = Auth::id();
 
         if ($validator->fails()) {
             return [
@@ -55,25 +56,41 @@ class CourseService
                 'errors' => $validator->errors(),
             ];
         }
-
-        // Handle file upload if needed
-
         $course->update($data);
 
         return [
             'success' => true,
             'message' => 'Course updated successfully',
-            'data' => $course,
+            'data' => new CourseResource($course),
         ];
+    }  
+
+    public function getAllCourses()
+    {
+        $course = Course::all();
+        return CourseResource::collection($course);
+
     }
 
-    public function deleteCourse(Course $course)
+    public function getCourseById($id)
     {
+        $course = Course::find($id);
+        if (!$course) {
+            abort(404, 'Course not found');
+        }
+        return $course;
+    }
+
+    public function deleteCourse(string $id): JsonResponse
+    {
+        $course = Course::findOrFail($id);
+
+        if (!$course) {
+            return response()->json(['message' => 'Course not found'], 404);
+        }
+
         $course->delete();
 
-        return [
-            'success' => true,
-            'message' => 'Course deleted successfully',
-        ];
+        return response()->json(['message' => 'Course deleted successfully']);
     }
 }
